split the public api into more files
